Extract GLState from context/mod.rs
